+++ plugins/postcss-lab-function/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased (minor)
 
-- Allow percentage units in more color components.
+- Allow percentage and number units in more color components.
 
 ```css
 .percentages {
@@ -16,6 +16,19 @@
 	color-2: rgb(181, 30, 19);
 	color-2: color(display-p3 0.65205 0.18193 0.12753);
 }
+
+.numbers {
+	color-1: lab(40 35 30);
+	color-2: lch(40 50 39);
+
+	/* becomes */
+
+	color-1: rgb(152, 68, 47);
+	color-1: color(display-p3 0.55453 0.28432 0.20788);
+	color-2: rgb(157, 63, 45);
+	color-2: color(display-p3 0.57072 0.27138 0.20109);
+}
+
 ```
 
 ### 4.1.1 (February 15, 2022)